+++ scripts/fix_report.py
@@ -1,1 +1,11 @@
-/Users/mpastorini/personal/pdm-conda/base/scripts/fix_report.py
+import os
+from pathlib import Path
+
+if __name__ == "__main__":
+    if (local_path := os.getenv("LOCAL_PATH", None)) is not None:
+        coverage_file = Path("coverage.xml")
+
+        with coverage_file.open("r+") as f:
+            contents = f.read().replace("/app/", f"{local_path}/")
+            f.seek(0)
+            f.write(contents)