+++ src/pdm_conda/__init__.py
@@ -8,7 +8,7 @@ if TYPE_CHECKING:
     from pdm.core import Core
 
 logger = termui.logger
-__version__ = "0.17.6"
+__version__ = "0.18.0"
 
 
 def main(core: Core):