+++ src/pdm_conda/installers/manager.py
@@ -1,11 +1,10 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, cast
+from typing import TYPE_CHECKING
 
 from pdm.installers import InstallManager
 
 from pdm_conda.conda import conda_install, conda_uninstall
-from pdm_conda.environments import CondaEnvironment
 from pdm_conda.models.candidates import CondaCandidate
 from pdm_conda.models.setup import CondaSetupDistribution
 
@@ -25,30 +24,31 @@ class CondaInstallManager(InstallManager):
         rename_pth: bool = False,
     ) -> None:
         super().__init__(environment, use_install_cache=use_install_cache, rename_pth=rename_pth)
-        self.environment = cast(CondaEnvironment, environment)
-        self._num_install = 0
-        self._num_remove = 0
-        self._batch_install: list[str] = []
-        self._batch_remove: list[str] = []
+        self._batch_install_queue: dict[str, str] = {}
+        self._batch_install_expected: set[str] = set()
+        self._batch_uninstall_queue: dict[str, str] = {}
+        self._batch_uninstall_expected: set[str] = set()
 
-    def prepare_batch_operations(self, num_install: int, num_remove: int):
-        self._num_install = num_install
-        self._num_remove = num_remove
+    def prepare_batch_operations(self, to_install: set[str], to_uninstall: set[str]):
+        self._batch_install_expected = to_install
+        self._batch_uninstall_expected = to_uninstall
 
-    def _run_with_conda(self, conda_func, new_requirement: str, requirements: list[str], min_requirements: int):
-        requirements.append(new_requirement)
-        if len(requirements) >= min_requirements:
-            try:
-                conda_func(
-                    self.environment.project,
-                    list(requirements),
-                    no_deps=True,
-                )
-                requirements.clear()
-            except:
-                if min_requirements == 0:
-                    requirements.clear()
-                raise
+    def _run_with_conda(
+        self,
+        conda_func,
+        new_requirement: str,
+        queue: dict[str, str],
+        expected: set[str],
+        op_name: str = "",
+    ):
+        if should_run := (new_requirement not in expected):
+            queue = {new_requirement: op_name or new_requirement}
+        else:
+            queue[new_requirement] = op_name or new_requirement
+            should_run = set(queue) == expected
+
+        if should_run:
+            conda_func(self.environment.project, list(queue.values()), no_deps=True)
 
     def install(self, candidate: Candidate) -> Distribution:
         """Install candidate, use conda if conda package else default installer.
@@ -58,9 +58,10 @@ class CondaInstallManager(InstallManager):
         if isinstance(candidate, CondaCandidate):
             self._run_with_conda(
                 conda_install,
+                candidate.name,
+                self._batch_install_queue,
+                self._batch_install_expected,
                 f"{candidate.link.url_without_fragment}#{candidate.link.hash}",
-                self._batch_install,
-                self._num_install,
             )
             return candidate.distribution
 
@@ -72,7 +73,12 @@ class CondaInstallManager(InstallManager):
         :param dist: distribution to uninstall
         """
         if isinstance(dist, CondaSetupDistribution):
-            self._run_with_conda(conda_uninstall, dist.name, self._batch_remove, self._num_remove)
+            self._run_with_conda(
+                conda_uninstall,
+                dist.name,
+                self._batch_uninstall_queue,
+                self._batch_uninstall_expected,
+            )
         else:
             super().uninstall(dist)